requirements.txt
streamlit>=1.31.0
pdfplumber>=0.10.2
langchain>=0.1.0
sentence-transformers>=2.2.2
groq>=0.4.0
easyocr>=1.7.0
Pillow>=9.0.0 